VariantVcfFactory: improve the memory usage by genotypes
Before:
===

with a file of 10K samples:
total memory: around 2.2 GB
900K instances (20%) of String (most of them are "0|0")
900K instances (20%) of char[] (most of them are "0|0")
char[] + String occupy 25% the total memory

After:
===
removed useless code and using String::intern

the memory usage was reduced to a 77% of the previous memory usage.

the intern method is used to reuse the string of the genotype
from the pool of strings.

However, there are still lots of maps that have different pointers
pointing to the same strings.

with a file of 10K samples:
total memory: around 1.7 GB
1M instances (32%) of TreeMap$Entry
1M instances (46%) of TreeMap
these two occupy 78% the total memory